clickhousex v0.1.0 Clickhousex

Clickhouse driver for Elixir.

This module handles the connection to Clickhouse, providing support for queries, connection backoff, logging, pooling and more.

Types

conn()

conn() :: DBConnection.conn()

A connection process name, pid or reference.

A connection reference is used when making multiple requests to the same connection, see transaction/3.

Functions

child_spec(opts)

child_spec(Keyword.t()) :: Supervisor.Spec.spec()

defaults(opts)

query(conn, statement, params, opts \\ [])

query(pid(), binary(), list(), Keyword.t()) :: {:ok, iodata(), Clickhousex.Result.t()}

query!(conn, statement, params, opts \\ [])

query!(pid(), binary(), list(), Keyword.t()) :: {iodata(), Clickhousex.Result.t()}

start_link(opts)

start_link(Keyword.t()) :: {:ok, pid()} | {:error, term()}

Connect to ClickHouse using ODBC. opts expects a keyword list with zero or more of:

* `:driver` - The driver the adapter will use.
    * default value: value of environment variable `CLICKHOUSE_ODBC_DRIVER`
* `:hostname` - The server hostname.
    * default value: localhost
* `:port` - The server port number.
    * default value: 8123
* `:database` - The name of the database.
    * default value: `default`
* `:username` - Username.
    * default value: empty
* `:password` - User's password.
    * default value: empty

timeout()